Move `swift_debug_settings.py` generation into Starlark (#1670)
Part of #1170.
