+++ examples/sanitizers/test/fixtures/bwb_targets_spec.json
@@ -41,9 +41,6 @@
             "_": "darwin_x86_64-dbg-STABLE-0/bin/test/fixtures/xcodeproj_bwb.generator-params/AddressSanitizerApp.0.link.params",
             "t": "g"
         },
-        "lldb_context": {
-            "c": "-iquote$(PROJECT_DIR) -iquote$(PROJECT_DIR)/bazel-out/ios-x86_64-min15.0-applebin_ios-ios_x86_64-dbg-STABLE-1/bin -O0 -DDEBUG=1 -fstack-protector -fstack-protector-all"
-        },
         "name": "AddressSanitizerApp",
         "outputs": {
             "p": true,
@@ -119,9 +116,6 @@
         "link_params": {
             "_": "darwin_x86_64-dbg-STABLE-0/bin/test/fixtures/xcodeproj_bwb.generator-params/ThreadSanitizerApp.1.link.params",
             "t": "g"
-        },
-        "lldb_context": {
-            "c": "-iquote$(PROJECT_DIR) -iquote$(PROJECT_DIR)/bazel-out/ios-x86_64-min15.0-applebin_ios-ios_x86_64-dbg-STABLE-1/bin -O0 -DDEBUG=1 -fstack-protector -fstack-protector-all"
         },
         "name": "ThreadSanitizerApp",
         "outputs": {
