+++ tools/generator/src/DTO/Target.swift
@@ -24,7 +24,6 @@ struct Target: Equatable {
     let appClips: Set<TargetID>
     var dependencies: Set<TargetID>
     var outputs: Outputs
-    let lldbContext: LLDBContext?
     let isUnfocusedDependency: Bool
     var additionalSchemeTargets: Set<TargetID>
 }
@@ -67,7 +66,6 @@ extension Target: Decodable {
         case appClips
         case dependencies
         case outputs
-        case lldbContext
         case isUnfocusedDependency
         case additionalSchemeTargets
     }
@@ -116,8 +114,6 @@ extension Target: Decodable {
         dependencies = try container.decodeTargetIDs(.dependencies)
         outputs = try container
             .decodeIfPresent(Outputs.self, forKey: .outputs) ?? .init()
-        lldbContext = try container
-            .decodeIfPresent(LLDBContext.self, forKey: .lldbContext)
         isUnfocusedDependency = try container
             .decodeIfPresent(Bool.self, forKey: .isUnfocusedDependency) ?? false
         additionalSchemeTargets = try container
