+++ tools/generator/src/Generator/CreateFilesAndGroups.swift
@@ -39,7 +39,6 @@ extension Generator {
     static let compileStubPath: Path = "_CompileStub_.m"
     static let externalFileListPath: Path = "external.xcfilelist"
     static let generatedFileListPath: Path = "generated.xcfilelist"
-    static let lldbSwiftSettingsModulePath: Path = "swift_debug_settings.py"
 
     private static let localizedGroupExtensions: Set<String> = [
         "intentdefinition",
@@ -622,144 +621,6 @@ already was set to `\(existingValue)`.
         addXCFileList(externalFileListPath, paths: externalPaths)
         addXCFileList(generatedFileListPath, paths: generatedPaths)
 
-        // Write target internal files
-
-        // - `lldbSwiftSettingsModule`
-
-        var lldbSettingsMap: [String: LLDBSettings] = [:]
-        for target in targets.values {
-            if let lldbContext = target.lldbContext,
-               let lldbSettingsKey = target.lldbSettingsKey
-            {
-                lldbSettingsMap[lldbSettingsKey] = LLDBSettings(
-                    frameworks: lldbContext.frameworkSearchPaths,
-                    includes: lldbContext.swiftmodules,
-                    clang: lldbContext.clang
-                )
-            }
-        }
-
-        let jsonEncoder = JSONEncoder()
-        jsonEncoder.keyEncodingStrategy = .convertToSnakeCase
-        jsonEncoder.outputFormatting = [
-            .prettyPrinted,
-            .sortedKeys,
-            .withoutEscapingSlashes,
-        ]
-        let lldbSettingsMapJSON = String(
-            data: try jsonEncoder.encode(lldbSettingsMap),
-            encoding: .utf8
-        )!
-
-        let lldbSwiftSettingsModule = #"""
-#!/usr/bin/python3
-
-"""An lldb module that registers a stop hook to set swift settings."""
-
-import lldb
-import re
-
-# Order matters, it needs to be from the most nested to the least
-_BUNDLE_EXTENSIONS = [
-    ".framework",
-    ".xctest",
-    ".appex",
-    ".bundle",
-    ".app",
-]
-
-_TRIPLE_MATCH = re.compile(r"([^-]+-[^-]+)(-\D+)[^-]*(-.*)?")
-
-_SETTINGS = \#(lldbSettingsMapJSON)
-
-def __lldb_init_module(debugger, _internal_dict):
-    # Register the stop hook when this module is loaded in lldb
-    ci = debugger.GetCommandInterpreter()
-    res = lldb.SBCommandReturnObject()
-    ci.HandleCommand(
-        "target stop-hook add -P swift_debug_settings.StopHook",
-        res,
-    )
-    if not res.Succeeded():
-        print(f"""\
-Failed to register Swift debug options stop hook:
-
-{res.GetError()}
-Please file a bug report here: \
-https://github.com/buildbuddy-io/rules_xcodeproj/issues/new?template=bug.md
-""")
-        return
-
-def _get_relative_executable_path(module):
-    for extension in _BUNDLE_EXTENSIONS:
-        prefix, _, suffix = module.rpartition(extension)
-        if prefix:
-            return prefix.split("/")[-1] + extension + suffix
-    return module.split("/")[-1]
-
-class StopHook:
-    "An lldb stop hook class, that sets swift settings for the current module."
-
-    def __init__(self, _target, _extra_args, _internal_dict):
-        pass
-
-    def handle_stop(self, exe_ctx, _stream):
-        "Method that is called when the user stops in lldb."
-        module = exe_ctx.frame.module
-        if not module:
-            return
-
-        module_name = module.file.__get_fullpath__()
-        versionless_triple = _TRIPLE_MATCH.sub(r"\1\2\3", module.GetTriple())
-        executable_path = _get_relative_executable_path(module_name)
-        key = f"{versionless_triple} {executable_path}"
-
-        settings = _SETTINGS.get(key)
-
-        if settings:
-            frameworks = " ".join([
-                f'"{path}"'
-                for path in settings["frameworks"]
-            ])
-            if frameworks:
-                lldb.debugger.HandleCommand(
-                    f"settings set -- target.swift-framework-search-paths {frameworks}",
-                )
-            else:
-                lldb.debugger.HandleCommand(
-                    "settings clear target.swift-framework-search-paths",
-                )
-
-            includes = " ".join([
-                f'"{path}"'
-                for path in settings["includes"]
-            ])
-            if includes:
-                lldb.debugger.HandleCommand(
-                    f"settings set -- target.swift-module-search-paths {includes}",
-                )
-            else:
-                lldb.debugger.HandleCommand(
-                    "settings clear target.swift-module-search-paths",
-                )
-
-            clang = settings["clang"]
-            if clang:
-                lldb.debugger.HandleCommand(
-                    f"settings set -- target.swift-extra-clang-flags '{clang}'",
-                )
-            else:
-                lldb.debugger.HandleCommand(
-                    "settings clear target.swift-extra-clang-flags",
-                )
-
-        return True
-
-"""#
-
-        files[.internal(lldbSwiftSettingsModulePath)] =
-            .nonReferencedContent(lldbSwiftSettingsModule)
-
         // Handle special groups
 
         rootElements.sortGroupedLocalizedStandard()
@@ -842,32 +703,6 @@ private struct LLDBSettings: Equatable, Encodable {
 // MARK: - Extensions
 
 private extension Target {
-    private func lldbSettingsKey(baseKey: String) -> String {
-        guard let productPath = product.path, product.type.isBundle else {
-            return baseKey
-        }
-
-        let executableName = product.executableName ??
-            productPath.path.lastComponentWithoutExtension
-
-        if platform.os == .macOS {
-            return "\(baseKey)/Contents/MacOS/\(executableName)"
-        } else {
-            return "\(baseKey)/\(executableName)"
-        }
-    }
-
-    var lldbSettingsKey: String? {
-        guard let productPath = product.path else {
-            return nil
-        }
-
-        let baseKey = """
-\(platform.versionlessTargetTriple) \(productPath.path.lastComponent)
-"""
-        return lldbSettingsKey(baseKey: baseKey)
-    }
-
     func xcodeSwiftGeneratedHeaderFilePath(_ filePath: FilePath) -> FilePath {
         guard let productPath = product.path else {
             // Should be caught earlier
