+++ xcodeproj/internal/platform.bzl
@@ -32,6 +32,15 @@ def _collect_platform(*, ctx):
         _platform = platform,
     )
 
+def _is_platform_type(platform, platform_type):
+    """Returns whether a platform is of a given type.
+
+    Args:
+        platform: A value returned from `platform_info.collect`.
+        platform_type: A value from `apple_common.platform`.
+    """
+    return platform._platform.platform_type == platform_type
+
 def _is_same_type(lhs, rhs):
     """Returns whether two platforms are the same platform type."""
     return lhs._platform.platform_type == rhs._platform.platform_type
@@ -53,6 +62,16 @@ def _platform_to_dto(platform):
     }
 
     return dto
+
+_LLDB_TRIPLE_PREFIX = {
+    apple_common.platform.ios_device: "ios",
+    apple_common.platform.ios_simulator: "ios",
+    apple_common.platform.macos: "macosx",
+    apple_common.platform.tvos_device: "tvos",
+    apple_common.platform.tvos_simulator: "tvos",
+    apple_common.platform.watchos_device: "watchos",
+    apple_common.platform.watchos_simulator: "watchos",
+}
 
 _SWIFT_TRIPLE_PREFIX = {
     apple_common.platform.ios_device: "ios",
@@ -86,9 +105,23 @@ def _platform_to_swift_triple(platform):
         triple_suffix = _TRIPLE_SUFFIX[platform._platform],
     )
 
+def _platform_to_lldb_context_triple(platform):
+    """Generates a lldb context triple for a platform.
+
+    Args:
+        platform: A value returned from `platform_info.collect`.
+    """
+    return "{arch}-apple-{triple_prefix}{triple_suffix}".format(
+        arch = platform._arch,
+        triple_prefix = _LLDB_TRIPLE_PREFIX[platform._platform],
+        triple_suffix = _TRIPLE_SUFFIX[platform._platform],
+    )
+
 platform_info = struct(
     collect = _collect_platform,
+    is_platform_type = _is_platform_type,
     is_same_type = _is_same_type,
     to_dto = _platform_to_dto,
+    to_lldb_context_triple = _platform_to_lldb_context_triple,
     to_swift_triple = _platform_to_swift_triple,
 )
